requirements.txt
Keras==2.2.2
web.py==0.40.dev1
matplotlib==2.2.2
tensorflow_gpu==1.5.0
opencv_contrib_python==3.4.2.17
numpy==1.15.0
Pillow==5.4.1
tensorflow==1.13.1
web==0.6.0
